uxceptional/pyproject.toml
[tool.poetry]
name = "uxceptional"
version = "0.1.0"
description = ""
authors = ["Noah Ablaseau <noaha@inmotionhosting.com>"]

[tool.poetry.dependencies]
python = "^3.10"
imgui = "^1.4.1"
glfw = "^2.5.3"
PyOpenGL = "^3.1.6"
Pillow = "^9.1.1"
ewmh = "^0.1.6"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
